employee

+++ global-service/src/main/java/in/cdac/university/globalService/entity/GbltConfigApplicationDataMst.java
@@ -87,5 +87,5 @@ public class GbltConfigApplicationDataMst implements Serializable {
 	private Integer unumCtypeId;
 
 	@Column(name = "unum_mtemple_id")
-	private Integer unumMtempleId;
+	private Long unumMtempleId;
 }
